+++ src/main.js
@@ -3,5 +3,5 @@
 
 require('./form.js');
 require('./game.js');
-require('./gallery-block/gallery.js');
+require('./gallery.js');
 require('./reviews-block/reviews.js');
